# contracts/openapi/bma-api-v1.yaml
openapi: 3.1.0
info:
  title: Binh Minh App API
  version: 1.0.0-alpha
  description: Versioned API for BMA mobile, Admin and source integrations.
servers:
  - url: https://gateway.abmtlab.com/bmapp/api/v1
paths:
  /auth/register:
    post:
      summary: Register a pending account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '202':
          description: Account pending approval
        '409':
          description: Username already exists
  /auth/login:
    post:
      summary: Login an approved account
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access and refresh token
        '401':
          description: Invalid credentials
        '403':
          description: Pending, rejected or disabled
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [username, password, display_name]
      properties:
        username: { type: string, minLength: 3, maxLength: 64 }
        password: { type: string, minLength: 10, maxLength: 200 }
        display_name: { type: string, minLength: 1, maxLength: 120 }
        employee_code: { type: [string, 'null'], maxLength: 64 }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
        device_name: { type: [string, 'null'], maxLength: 100 }
...